brahma_hq_reviewer/requirements.txt

fastapi
uvicorn[standard]
google-cloud-storage
jinja2
python-multipart
